save to filesystem, don't store in memory
Summary
---

We were using up a lot of memory with the old implementation.
Since we're just saving them to disk in the AssetBuild process,
let's save them to disk in the first place.

How?
---

Stop using `webpack-dev-middleware`. Use the standard `webpack`
API which can `watch()` for changes. This writes the bundles to
disk.

Still include the web API, but only use it for starting a bundle,
reloading a config, and checking if a bundle is completed yet.