watcher: fix subtle-ass bug
When an ES6 arrow function is surrounded by brackets, it's treated
as a block and you must explicitly return a value. When it's not,
it's treated as an expression and the value of the expression is
returned.

Returning null here made a further call to Promise.race() resolve
successfully immediately (seems like wacky behavior).

From the Promise.race spec:
  > If the iterable contains one or more non-promise value and/or an
    already resolved/rejected promise, then Promise.race will
    resolve to the first of these values found in the iterable.

So since null is a non-promise value, it resolves immediately to that.