Compiler Expiration: clear on reload
Fix a bug where we wouldn't clear the expire timeout when replacing
the webpack compiler. Thus if you reloaded, {timeout} seconds later
it would kill the compiler, seemingly out of no-where.

Now we clear the timeout before replacing (in set) and before
removing it.

I've decided to not store the expiration timer on the compiler,
but connected to the username. There is now always only one timer
and one compiler.