Builds: save and pass along last build error
If the webpack wasn't currently building, we'd assume it built
successfully. The affect was this pattern:

Webpack: building ...
Client Req1: > Is the build ready? -> blocking...
Webpack: finished! and failed!
Client Req1: < Finished and failed!

Client Req2: > Is the build ready?
Client Req2: < Finished and Succeeded!

We ended up lying about the success state of the build.

Now, we store the resultant error after every build finishes
and send that along each time a client asks.