Use the user's `gid` for the new process's `gid`
Without this, the process isn't allowed to read or write files which
the user's group is allowed to read or write.

We ran into a problem where the parent directory of our webpack output
directory only had write permissions for `apache` and `coders`, and
the process (unlike all our users) didn't have the `coders` group, so
it couldn't write the directory.